fix(dashboard): keep React out of App's compile via FilterChipDropdown types
compile-app has been failing on master with:

  FeatureSet/Dashboard/src/Components/ResourceOwners/FilterChipDropdown.tsx(11,8):
  error TS2307: Cannot find module 'react' or its corresponding type declarations.

App's tsconfig excludes FeatureSet/Dashboard, but `exclude` only trims the
root file list — an excluded file is still type-checked when something in the
program imports it. App/Tests/Dashboard/FacetSelectionState.test.ts imports
FacetSelectionState, which imported FilterOperator from the FilterChipDropdown
component, so the whole React component was pulled into App's program. The
compile-app job only runs `cd App && npm install`, and React is a dependency of
the Dashboard package, not of App.

Move the React-free half of the module — FilterChipDropdownOption,
FilterOperator, FILTER_OPERATOR_LABELS, isValueOperator — into
FilterChipDropdownTypes.ts and point FacetSelectionState at it. The component
re-exports all four, so every existing `from "./FilterChipDropdown"` import
keeps working.

Verified: `cd App && npx tsc --noEmit` is clean without the Dashboard package's
node_modules present (which is what CI does), the Dashboard package still
compiles, and Tests/Dashboard/FacetSelectionState.test.ts passes.
